+++ packages/react/src/components/Dropdown/DropdownSelectedItem.tsx
@@ -139,10 +139,11 @@ class DropdownSelectedItem extends UIComponent<WithAsProp<DropdownSelectedItemPr
         styles: styles.header,
       }),
     })
-    const renderIcon = _.isNil(icon)
+    const iconProps = _.isNil(icon)
       ? icon
-      : render =>
-          render(icon, (ComponentType, props) =>
+      : {
+          name: null,
+          children: (ComponentType, props) =>
             Icon.create(icon, {
               defaultProps: () => ({
                 'aria-label': `Remove ${header} from selection.`, // TODO: Extract this in a behaviour.
@@ -151,11 +152,11 @@ class DropdownSelectedItem extends UIComponent<WithAsProp<DropdownSelectedItemPr
               }),
               overrideProps: this.handleIconOverrides(props),
             }),
-          )
-    const renderImage = _.isNil(image)
+        }
+    const imageProps = _.isNil(image)
       ? image
-      : render =>
-          render(image, (ComponentType, props) =>
+      : {
+          children: (ComponentType, props) =>
             Image.create(image, {
               defaultProps: () => ({
                 avatar: true,
@@ -164,7 +165,7 @@ class DropdownSelectedItem extends UIComponent<WithAsProp<DropdownSelectedItemPr
               }),
               overrideProps: props,
             }),
-          )
+        }
 
     return (
       <Ref innerRef={this.itemRef}>
@@ -176,8 +177,8 @@ class DropdownSelectedItem extends UIComponent<WithAsProp<DropdownSelectedItemPr
           onClick={this.handleClick}
           onKeyDown={this.handleKeyDown}
           content={contentElement}
-          icon={renderIcon}
-          image={renderImage}
+          icon={iconProps}
+          image={imageProps}
           {...unhandledProps}
         />
       </Ref>